requirements.txt
pandas>=1.3.0
sec-edgar-downloader>=5.0.0
beautifulsoup4>=4.9.0
requests>=2.25.0
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-cov>=4.1.0
PyYAML>=6.0
